refactor(ui): reorganize frontend structure
- routes/ is now flat — one file per view, no subdirectories
- Shared xyflow building blocks moved to components/flow/:
    AppNode, CapabilityNode, FlowControls
- Remove dead code: GraphView.svelte (only served integration-map,
  now removed) and lib/cytoscape.js (no longer used by any view)
- Remove generic /view/:viewName/graph route from App.svelte (was only
  used by the now-removed integration-map view)
- Update all import paths accordingly
