+++ cmd/archipulse/ui/src/components/flow/AppNode.svelte
