+++ cmd/archipulse/ui/src/components/flow/CapabilityNode.svelte
