+++ cmd/archipulse/ui/src/routes/CapabilityTree.svelte
@@ -11,9 +11,9 @@
   import '@xyflow/svelte/dist/style.css';
   import dagre from '@dagrejs/dagre';
   import { api } from '../lib/api.js';
-  import CapabilityNode from './capability/CapabilityNode.svelte';
-  import AppNode       from './dependency/AppNode.svelte';
-  import FlowControls  from './dependency/FlowControls.svelte';
+  import CapabilityNode from '../components/flow/CapabilityNode.svelte';
+  import AppNode       from '../components/flow/AppNode.svelte';
+  import FlowControls  from '../components/flow/FlowControls.svelte';
 
   let { params = {} } = $props();
